refactor: split Enum runtime functions, split tests
