+++ .eslintrc.json
@@ -9,6 +9,7 @@
 	],
 	"ignorePatterns": ["dist/**"],
 	"rules": {
-		"@typescript-eslint/no-namespace": "off"
+		"@typescript-eslint/no-namespace": "off",
+		"@typescript-eslint/no-explicit-any": "off"
 	}
 }
