+++ src/enum/is.spec.ts
@@ -1,89 +1,5 @@
-import { Enum } from "./enum.js";
-import type { Equal, Expect } from "./shared/tests.js";
-
-describe("Enum", () => {
-	test("Default", () => {
-		type Event = Enum<{
-			Open: true;
-			Data: { value: unknown };
-			Close: true;
-		}>;
-
-		const Event = Enum({} as Event);
-
-		{
-			const eventOpen = Event.Open();
-			expect(eventOpen).toStrictEqual({ _type: "Open" });
-			const eventData = Event.Data({ value: "..." });
-			expect(eventData).toStrictEqual({ _type: "Data", value: "..." });
-			const eventClose = Event.Close();
-			expect(eventClose).toStrictEqual({ _type: "Close" });
-
-			void [eventOpen, eventData, eventClose];
-		}
-	});
-
-	test("Mapped", () => {
-		type Event = Enum<{
-			Open: true;
-			Data: { value: unknown };
-			Close: true;
-		}>;
-
-		const Event = Enum({} as Event, {
-			Data: (value: unknown) => ({ value }),
-		});
-
-		{
-			const eventOpen = Event.Open();
-			expect(eventOpen).toStrictEqual({ _type: "Open" });
-			const eventData = Event.Data("...");
-			expect(eventData).toStrictEqual({ _type: "Data", value: "..." });
-			const eventClose = Event.Close();
-			expect(eventClose).toStrictEqual({ _type: "Close" });
-
-			void [eventOpen, eventData, eventClose];
-		}
-	});
-
-	test("Custom", () => {
-		type Event = Enum<
-			{
-				Open: true;
-				Data: { value: unknown };
-				Close: true;
-			},
-			"custom"
-		>;
-
-		const Event = Enum(
-			{} as Event,
-			{
-				Data: (value: unknown) => ({ value }),
-			},
-			"custom",
-		);
-
-		{
-			const eventOpen: Event = { custom: "Open" };
-			const eventData: Event = { custom: "Data", value: "..." };
-			const eventClose: Event = { custom: "Close" };
-
-			void [eventOpen, eventData, eventClose];
-		}
-
-		{
-			const eventOpen = Event.Open();
-			expect(eventOpen).toStrictEqual({ custom: "Open" });
-			const eventData = Event.Data("...");
-			expect(eventData).toStrictEqual({ custom: "Data", value: "..." });
-			const eventClose = Event.Close();
-			expect(eventClose).toStrictEqual({ custom: "Close" });
-
-			void [eventOpen, eventData, eventClose];
-		}
-	});
-});
+import type { Equal, Expect } from "../shared/tests.js";
+import { Enum } from "../enum.js";
 
 describe("Enum.is", () => {
 	test("Default", () => {
