+++ src/enum/match.spec.ts
@@ -1,6 +1,5 @@
-import type { Expect, Equal } from "./shared/tests.js";
-import type { Enum } from "./enum.js";
-import { match } from "./match.js";
+import { Enum } from "../enum.js";
+import type { Expect, Equal } from "../shared/tests.js";
 
 describe("match", () => {
 	describe("default discriminant", () => {
@@ -16,7 +15,7 @@ describe("match", () => {
 			// all cases
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: "One" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "One">>;
@@ -25,7 +24,7 @@ describe("match", () => {
 			// all cases as fns
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "One">>;
@@ -34,7 +33,7 @@ describe("match", () => {
 			// only fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						_: () => "Fallback" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "Fallback">>;
@@ -56,7 +55,7 @@ describe("match", () => {
 			// all cases
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: "One" as const,
 						Two: "Two" as const,
 					});
@@ -66,7 +65,7 @@ describe("match", () => {
 			// all cases as fns
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 						Two: () => "Two" as const,
 					});
@@ -76,7 +75,7 @@ describe("match", () => {
 			// only fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						_: "Fallback" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "Fallback">>;
@@ -85,7 +84,7 @@ describe("match", () => {
 			// only fn fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						_: () => "Fallback" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "Fallback">>;
@@ -94,7 +93,7 @@ describe("match", () => {
 			// one case and fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 						_: "Fallback" as const,
 					});
@@ -104,7 +103,7 @@ describe("match", () => {
 			// all cases and fn fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 						_: () => "Fallback" as const,
 					});
@@ -114,7 +113,7 @@ describe("match", () => {
 			// all cases and undefined as fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 						_: undefined,
 					});
@@ -124,7 +123,7 @@ describe("match", () => {
 			// all cases, some using value properties
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => 0,
 						Two: ({ value }) => value,
 						_: () => "Unknown",
@@ -139,7 +138,7 @@ describe("match", () => {
 			// all cases
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 						Two: () => "Two" as const,
 						_: undefined,
@@ -151,7 +150,7 @@ describe("match", () => {
 
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						_: () => "Unknown",
 					});
 
@@ -162,7 +161,7 @@ describe("match", () => {
 
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => 0,
 						Two: ({ value }) => value,
 						_: () => "Unknown",
@@ -188,7 +187,7 @@ describe("match", () => {
 			// all cases
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "One">>;
@@ -197,7 +196,7 @@ describe("match", () => {
 			// only fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						_: () => "Fallback" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "Fallback">>;
@@ -206,7 +205,7 @@ describe("match", () => {
 			// all cases and forbid fallback
 			{
 				const fn = (value: Value) =>
-					match(value, {
+					Enum.match(value, {
 						One: () => "One" as const,
 						_: undefined,
 					});
@@ -229,7 +228,7 @@ describe("match", () => {
 			// all cases
 			{
 				const fn = (value: Value) =>
-					match([value, "custom"], {
+					Enum.match([value, "custom"], {
 						One: () => "One" as const,
 						Two: () => "Two" as const,
 					});
@@ -239,7 +238,7 @@ describe("match", () => {
 			// only fallback
 			{
 				const fn = (value: Value) =>
-					match([value, "custom"], {
+					Enum.match([value, "custom"], {
 						_: () => "Fallback" as const,
 					});
 				!0 as Expect<Equal<ReturnType<typeof fn>, "Fallback">>;
@@ -248,7 +247,7 @@ describe("match", () => {
 			// one case and fallback
 			{
 				const fn = (value: Value) =>
-					match([value, "custom"], {
+					Enum.match([value, "custom"], {
 						One: () => "One" as const,
 						_: () => "Fallback" as const,
 					});
@@ -258,7 +257,7 @@ describe("match", () => {
 			// all cases and fallback
 			{
 				const fn = (value: Value) =>
-					match(
+					Enum.match(
 						value,
 						{
 							One: () => "One" as const,
@@ -273,7 +272,7 @@ describe("match", () => {
 			// all cases, some using value properties
 			{
 				const fn = (value: Value) =>
-					match(
+					Enum.match(
 						value,
 						{
 							One: () => 0,
@@ -292,7 +291,7 @@ describe("match", () => {
 			// all cases
 			{
 				const fn = (value: Value) =>
-					match(
+					Enum.match(
 						value,
 						{
 							One: () => "One" as const,
@@ -308,7 +307,7 @@ describe("match", () => {
 
 			{
 				const fn = (value: Value) =>
-					match([value, "custom"], {
+					Enum.match([value, "custom"], {
 						_: () => "Unknown",
 					});
 
@@ -319,7 +318,7 @@ describe("match", () => {
 
 			{
 				const fn = (value: Value) =>
-					match(
+					Enum.match(
 						value,
 						{
 							One: () => 0,
